campus360/backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
django-cors-headers>=4.3
django-filter>=23.5
djangorestframework-simplejwt>=5.3
psycopg2-binary>=2.9
python-dotenv>=1.0
gunicorn>=21.2
whitenoise>=6.6
supabase>=2.0
Pillow>=10.0
razorpay>=1.4
qrcode[pil]>=7.4
